requirements.txt
blinker==1.8.2
click==8.1.7
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
jinja2==3.1.4
markupsafe==2.1.5
numpy==2.0.1
pandas==2.2.2
psycopg2-binary==2.9.9
pytz==2024.1
sqlalchemy==2.0.31
tzdata==2024.1
werkzeug==3.0.3
